[TASK] Replace $_EXTKEY and remove from ext_tables
It is recommended to check for the existence of the constants
defined('TYPO3_MODE') or die(); at the top of ext_tables.php and
ext_localconf.php files to make sure the file is executed only
indirectly within TYPO3 context. This is a security measure since this
code in global scope should not be executed through the web server
directly as entry point.

Additionally, it is recommended to use the extension name (e.g.
“tt_address”) instead of $_EXTKEY within the two configuration files as
this variable will be removed in the future. This also applies to
$_EXTCONF.

However, due to limitations to TER, the $_EXTKEY option should be kept
within an extension’s ext_emconf.php.