Use UserFileMountService of core
Instead of using an outdated copy of the
UserFileMountService from core simply override
the config and use the core service.

The UserFileMountService is deprecated in v12 and will be removed in v13, but currently selection of a storage root folder is not possible from the element browser, therefore migration to type=folder is not an option currently.

https://forge.typo3.org/issues/100789