[BUGFIX] Respect groups set by ModifyResolvedFrontendGroupsEvent
The Core provides an event `ModifyResolvedFrontendGroupsEvent` to
add groups to the current user. This event can be used regardless
of the login status of the user, so that groups can also be added to
an anonymous user session. This is what some extensions do, e. g.
b13/sessionpassword.

This patch adjusts the permission check to also check groups from
anonymous user sessions.