[TASK] Update for TYPO3 v13 (#259)
* [TASK] Update securedownload for TYPO3 v13

* [TASK] Set permission rights for backend user

* [TASK] Rector, Fractor, phpcs

* [TASK] add local dev tools

* [TASK] use dedicated function for UC unpack

* [BUGFIX] PHP warnings and missing variable

* [BUGFIX] Respect groups set by ModifyResolvedFrontendGroupsEvent

The Core provides an event `ModifyResolvedFrontendGroupsEvent` to
add groups to the current user. This event can be used regardless
of the login status of the user, so that groups can also be added to
an anonymous user session. This is what some extensions do, e. g.
b13/sessionpassword.

This patch adjusts the permission check to also check groups from
anonymous user sessions.

* [TASK] Update composer.json infos

* [TASK] Simplify VH code

Override verdict-method and remove useless render method.

* [BUGFIX] add list_type to CType mapping to Upgrade Wizard

* [DOC] Update README.md

* [TASK] use thecodingmachine/php for php tools

* [BUGFIX] replace call to non-existing function getCountExpression with COUNT()

---------

Co-authored-by: Johannes Böttcher <johannes.boettcher@vancado.de>
Co-authored-by: Simon Praetorius <moin@praetorius.me>
Co-authored-by: Tobias Kapplmüller <47564597+siwa-tkapplmueller@users.noreply.github.com>
Co-authored-by: siwa_tkapplmueller <t.kapplmueller@siwa.at>